// src/core/boc/cell-type.ts
/**
 * Copyright (c) Whales Corp.
 * All Rights Reserved.
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 */

export enum CellType {
    Ordinary = -1,
    PrunedBranch = 1,
    Library = 2,
    MerkleProof = 3,
    MerkleUpdate = 4,
}